Sending player data correctly

+++ server/src/main/java/com/browsergame/server/game/Player.java
@@ -3,25 +3,25 @@ package com.browsergame.server.game;
 public class Player
 {
     private final String id;
-    private float x;
-    private float y;
+    private int x;
+    private int y;
 
     public Player(String playerId)
     {
         id = playerId;
     }
 
-    public float getX()
+    public int getX()
     {
         return x;
     }
 
-    public float getY()
+    public int getY()
     {
         return y;
     }
 
-    public void setPos(float newX, float newY)
+    public void setPos(int newX, int newY)
     {
         x = newX;
         y = newY;
